+++ mvd-reader/events/events.go
@@ -66,6 +66,7 @@ type (
 	BackpackDropHintEvent    = parser.BackpackDropHintEvent
 	ItemPickupHintEvent      = parser.ItemPickupHintEvent
 	BackpackPickupHintEvent  = parser.BackpackPickupHintEvent
+	DemoMarkEvent            = parser.DemoMarkEvent
 	ItemPickupPrintEvent     = parser.ItemPickupPrintEvent
 	BackpackPickupPrintEvent = parser.BackpackPickupPrintEvent
 	DemoStartTimestampEvent  = parser.DemoStartTimestampEvent